notify user, if entities are missing (i.e. the user forgot to add them to the project file)
Xilinx considers this as a warning only, but it will cause lots of errors in the simulation (UUT is not ready because the ready signal is 'U'). IMHO, the error message don't direct the user in the right direction, so we make sure that the user fixes that kind of problem before we start the simulation.