bugfix for PrettyPrintBuilder
Seperator will be missing after the first item, if the first item
is a container. I think it will also be missing before containers
(but I haven't confirmed that). This occurs because the container
is empty when it is added, so the builder thinks that it shouldn't
add the seperator.

fix: defer adding the container until we know whether it will be
empty or not