only run synthesis, if the bitstream file doesn't exist
This breaks incremental build because the task won't run, if an input
changes. However, this didn't work well, anyway. This change means
that the user won't have to wait for needless synthesis runs to
finish, but it also means that she must be careful to always clean
the output, if she changes any input (directly or indirectly).